+++ CHANGELOG.md
@@ -191,6 +191,7 @@
 
 ### Minor changes
 
+* Print CLICON_YANG_DIR and CLICON_FEATURE lists on startup with debug flag 
 * Extended `util/clixon_util_xml` with yang and validate functionality so it can be used as a stand-alone utility for validating XML/JSON files
 * JSON parse and print improvements
   * Integrated parsing with namespace translation and yang spec lookup